chore(simple-form/layout.tsx): move sonner import to pages/layout in attempt to shake from bundle
