+++ next-portfolio/components/index.ts
@@ -22,7 +22,6 @@ export * from "./resizable";
 export * from "./select";
 export { default as StyledGoBack } from "./styled-go-back";
 export { default as Textarea } from "./textarea";
-export { default as Toaster } from "./toast";
 export { default as ViewSource } from "./view-source";
 
 export * from "./form";
